+++ blocks/mdaqenc.h
@@ -2,6 +2,6 @@
 #define _MDAQENC_H
 
 void ENCInit(unsigned char Channel);
-void ENCStep(unsigned char Channel, int *position);
+void ENCStep(unsigned char Channel, long *position);
 
 #endif /* _MDAQENC_H */
